+++ internal/controllers/order-item.go
@@ -53,7 +53,7 @@ func (c *OrderItemController) Get(ctx echo.Context) error {
 		return utils.ErrorResponse(ctx, http.StatusBadRequest, "Invalid ID")
 	}
 
-	orderItem, err := c.service.Get(ctx.Request().Context(), uint(id))
+	orderItem, err := c.service.Get(ctx.Request().Context(), id)
 	if err != nil {
 		return utils.ErrorResponse(ctx, http.StatusNotFound, "Order item not found")
 	}
@@ -86,7 +86,7 @@ func (c *OrderItemController) Delete(ctx echo.Context) error {
 		return utils.ErrorResponse(ctx, http.StatusBadRequest, "Invalid ID")
 	}
 
-	if err := c.service.Delete(ctx.Request().Context(), uint(id)); err != nil {
+	if err := c.service.Delete(ctx.Request().Context(), id); err != nil {
 		return utils.ErrorResponse(ctx, http.StatusInternalServerError, "Failed to delete order item")
 	}
